+++ lib/legacy_screenutil_scope.dart
@@ -1,0 +1,3 @@
+library screen_adapt_legacy_screenutil_scope_api;
+
+export 'src/widgets/legacy_screenutil_scope.dart';